Add hilt module for dependency injection Modules for the databases. Should be changed with repositories
