+++ app/src/main/java/com/github/sdp/ratemyepfl/database/CourseDatabase.kt
@@ -4,6 +4,5 @@ import com.github.sdp.ratemyepfl.model.items.Course
 import com.github.sdp.ratemyepfl.model.review.CourseReview
 
 interface CourseDatabase {
-    fun getReviews(): List<CourseReview>
     fun getCourses(): List<Course>
 }
