+++ app/src/main/java/com/github/sdp/ratemyepfl/database/CourseReviewDatabase.kt
@@ -1,0 +1,7 @@
+package com.github.sdp.ratemyepfl.database
+
+import com.github.sdp.ratemyepfl.model.review.CourseReview
+
+interface CourseReviewDatabase {
+    fun getReviews(): List<CourseReview>
+}
