+++ app/src/main/java/com/github/sdp/ratemyepfl/dependencyinjection/CourseDatabaseModule.kt
@@ -1,5 +1,6 @@
 package com.github.sdp.ratemyepfl.dependencyinjection
 
+import android.app.Application
 import com.github.sdp.ratemyepfl.database.CourseDatabase
 import com.github.sdp.ratemyepfl.database.FakeCoursesDatabase
 import dagger.Binds
@@ -7,10 +8,12 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewComponent
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
 
 @Module
-@InstallIn(ViewComponent::class)
+@InstallIn(SingletonComponent::class)
 abstract class CourseDatabaseModule {
-    @Provides
-    fun providesCourseDatabase(): CourseDatabase = FakeCoursesDatabase()
+    @Binds
+    abstract fun bindsCourseDatabase(db: FakeCoursesDatabase): CourseDatabase
 }
