+++ app/src/main/java/com/github/sdp/ratemyepfl/viewmodel/ReviewListViewModel.kt
@@ -2,13 +2,15 @@ package com.github.sdp.ratemyepfl.viewmodel
 
 import androidx.lifecycle.ViewModel
 import com.github.sdp.ratemyepfl.database.CourseDatabase
+import com.github.sdp.ratemyepfl.database.CourseReviewDatabase
 import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
 
 /**
  * TO DO: change the database to repository, when available
  */
 @HiltViewModel
-class ReviewListViewModel(val database: CourseDatabase) : ViewModel() {
+class ReviewListViewModel @Inject constructor(val database: CourseReviewDatabase) : ViewModel() {
 
     fun getReviews() = database.getReviews()
